separates board into two modules, and small refactor to Board.generate_blank_board
