+++ lib/command_line_io.ex
@@ -1,6 +1,7 @@
 defmodule CommandLineIO do
   import String, only: [rstrip: 1]
   @rules Rules
+  @presenter BoardPresenter
 
   def write(msg) do
     IO.write(msg)
@@ -15,7 +16,7 @@ defmodule CommandLineIO do
   end
 
   def display_board do
-    write("Here's what the board looks like:\n#{TicTacToeBoard.display_sample_board}\n")
+    write("Here's what the board looks like:\n#{@presenter.display_sample_board}\n")
   end
 
   def get_first_or_second do
@@ -38,12 +39,12 @@ defmodule CommandLineIO do
     write("Bye!\n")
   end
 
-  def display_current_board(board, updated_board) do
-    write("\nCurrent Board:\n#{board.display(updated_board)}")
+  def display_current_board(updated_board) do
+    write("\nCurrent Board:\n#{@presenter.display(updated_board)}")
   end
 
-  def display_game_over_message(board, updated_board_state) do
-    write("\nCurrent Board:\n#{board.display(updated_board_state)}\n#{create_game_over_message(updated_board_state)}")
+  def display_game_over_message(updated_board_state) do
+    write("\nCurrent Board:\n#{@presenter.display(updated_board_state)}\n#{create_game_over_message(updated_board_state)}")
   end
 
   def create_game_over_message(current_board) do
