+++ test/board_test.exs
@@ -8,25 +8,6 @@ defmodule TicTacToeBoardTest do
                                           "", "", ""]
   end
 
-  test "#display_sample_board: returns a sample board for the beginning of the game" do
-    assert @board.display_sample_board == " 1 | 2 | 3 \n" <>
-                                       " ___________ \n" <>
-                                        " 4 | 5 | 6 \n" <>
-                                       " ___________ \n" <>
-                                         " 7 | 8 | 9 "
-  end
-
-  test "#display: returns the current board for display" do
-    current_board = ["", "", "",
-                     "X", "", "",
-                     "", "", ""]
-    assert @board.display(current_board) == "  |  |  \n" <>
-                                            " ___________ \n" <>
-                                            " X |  |  \n" <>
-                                            " ___________ \n" <>
-                                            "  |  |  \n"
-  end
-
   test "#update: updates the board with the given mark at the given position" do
      current_board = ["", "", "",
                       "X", "", "",
